+++ src/main/kotlin/kr/co/imoscloud/entity/inventory/InventoryIn.kt
@@ -28,7 +28,7 @@ class InventoryIn {
     var inType: String? = null
 
     @Column(name = "QTY")
-    var qty: Int? = null
+    var qty: Double? = null
 
     @Column(name = "UNIT_PRICE")
     var unitPrice: Int? = null
